Separate current Md5Cache handling
