+++ lib/hard-context-module-factory-plugin.js
@@ -35,7 +35,7 @@ class HardContextModuleFactoryPlugin {
           moduleCache,
           moduleResolveCache,
           moduleResolveCacheChange,
-        } = caches();
+        } = caches(compilation);
 
         var hardContextFactory = new HardContextModuleFactory({
           compilation: compilation,
